python-scanner/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-nmap==0.7.1
requests==2.31.0
pydantic==2.5.0